fashion-mnist/project/requirements.txt
opencv-python
PILLOW
requests
numpy
boto3
